Integration test for clean shutdown
Bring up server in a dedicated thread and check that server code does
not indicate an error, when it is brought down by SIGINT sent to the
dedicated thread.

Make state indicate when server is up and listeninig for new
connections. In test code, this is used to check that thread running
server should be blocking. While blocking, user or a program may decide
to terminate server. Test that termination by SIGINT causes the server
to be brought down in a controlled way and without indicated errors.